requirements.txt
agntcy-app-sdk>=0.4.1
a2a>=0.3.0
langgraph>=0.4.1
langchain-core
litellm
flask>=3.0
uvicorn
requests>=2.31
beautifulsoup4>=4.12
lxml>=5.0
scikit-learn>=1.3
anthropic>=0.40
openai>=1.0
python-dotenv>=1.0
pydantic>=2.0
# Observability (ioa-observe) is handled via docker-compose services, not pip
playwright>=1.40
